backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.2.1